MaximumMoveListDistanceCalculator
