+++ src/harness/testcases/cu_pass/features/steps/dpa_neighborhood/neighborhood_calculation.py
@@ -2,7 +2,7 @@ from math import inf
 
 from behave import *
 
-from cu_pass.dpa_calculator.aggregate_interference_calculator.aggregate_interference_calculator_winnforum import \
+from cu_pass.dpa_calculator.aggregate_interference_calculator.aggregate_interference_calculator_winnforum.support.maximum_move_list_distance_calculator import \
     WINNFORUM_MINIMUM_INTERFERENCE
 from cu_pass.dpa_calculator.aggregate_interference_calculator.aggregate_interference_monte_carlo_calculator.aggregate_interference_monte_carlo_calculator import \
     AggregateInterferenceMonteCarloCalculator, AggregateInterferenceMonteCarloResults, AggregateInterferenceTypes, \
